# OpenAPI/threat-actors.openapi.json
{"openapi":"3.1.0","info":{"title":"TIP Threat Actors","description":"Threat actor library — MITRE ATT&CK, ransomware groups, tools and TTPs","version":"0.1.0"},"paths":{"/health":{"get":{"tags":["meta"],"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object","title":"Response Health Health Get"}}}}}}},"/actors":{"get":{"tags":["actors"],"summary":"List Actors","operationId":"list_actors_actors_get","parameters":[{"name":"sector","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Sector"}},{"name":"country","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Country"}},{"name":"motivation","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Motivation"}},{"name":"status","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Status"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":500,"default":100,"title":"Limit"}},{"name":"offset","in":"query","required":false,"schema":{"type":"integer","minimum":0,"default":0,"title":"Offset"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ActorOut"},"title":"Response List Actors Actors Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/actors/{actor_id}":{"get":{"tags":["actors"],"summary":"Get Actor","operationId":"get_actor_actors__actor_id__get","parameters":[{"name":"actor_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Actor Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ActorDetailOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/actors/{actor_id}/insight":{"get":{"tags":["actors"],"summary":"Get Actor Insight","operationId":"get_actor_insight_actors__actor_id__insight_get","parameters":[{"name":"actor_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Actor Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ActorInsightOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/actors/{actor_id}/ttps":{"get":{"tags":["actors"],"summary":"List Actor Ttps","operationId":"list_actor_ttps_actors__actor_id__ttps_get","parameters":[{"name":"actor_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Actor Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{},"title":"Response List Actor Ttps Actors  Actor Id  Ttps Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/ransomware/groups":{"get":{"tags":["ransomware"],"summary":"List Groups","operationId":"list_groups_ransomware_groups_get","parameters":[{"name":"status","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Status"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":500,"default":100,"title":"Limit"}},{"name":"offset","in":"query","required":false,"schema":{"type":"integer","minimum":0,"default":0,"title":"Offset"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RansomwareGroupOut"},"title":"Response List Groups Ransomware Groups Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/ransomware/groups/{group_id}":{"get":{"tags":["ransomware"],"summary":"Get Group","operationId":"get_group_ransomware_groups__group_id__get","parameters":[{"name":"group_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Group Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RansomwareGroupOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/ransomware/victims":{"get":{"tags":["ransomware"],"summary":"List Victims","operationId":"list_victims_ransomware_victims_get","parameters":[{"name":"group_id","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"Group Id"}},{"name":"since","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Since"}},{"name":"country","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Country"}},{"name":"sector","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Sector"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":500,"default":100,"title":"Limit"}},{"name":"offset","in":"query","required":false,"schema":{"type":"integer","minimum":0,"default":0,"title":"Offset"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RansomwareVictimOut"},"title":"Response List Victims Ransomware Victims Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/tools":{"get":{"tags":["tools"],"summary":"List Tools","operationId":"list_tools_tools_get","parameters":[{"name":"type","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"malware or tool","title":"Type"},"description":"malware or tool"},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":500,"default":100,"title":"Limit"}},{"name":"offset","in":"query","required":false,"schema":{"type":"integer","minimum":0,"default":0,"title":"Offset"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ToolOut"},"title":"Response List Tools Tools Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/tools/{tool_id}":{"get":{"tags":["tools"],"summary":"Get Tool","operationId":"get_tool_tools__tool_id__get","parameters":[{"name":"tool_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Tool Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ToolOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/ttps/{technique_id}":{"get":{"tags":["tools"],"summary":"Actors Using Technique","description":"Returns actors that use the given ATT&CK technique.","operationId":"actors_using_technique_ttps__technique_id__get","parameters":[{"name":"technique_id","in":"path","required":true,"schema":{"type":"string","title":"Technique Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/refresh":{"post":{"tags":["ingest"],"summary":"Trigger Refresh","description":"Scheduler trigger: incremental refresh (MITRE + groups + recent victims).","operationId":"trigger_refresh_refresh_post","responses":{"202":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/refresh/full":{"post":{"tags":["ingest"],"summary":"Trigger Full Refresh","description":"First-boot / reconciliation trigger: full historical pull (~80k victims).","operationId":"trigger_full_refresh_refresh_full_post","responses":{"202":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health/sources":{"get":{"tags":["health"],"summary":"Source Health","operationId":"source_health_health_sources_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/SourceHealthOut"},"type":"array","title":"Response Source Health Health Sources Get"}}}}}}}},"components":{"schemas":{"ActorDetailOut":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"mitre_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Mitre Id"},"name":{"type":"string","title":"Name"},"aliases":{"items":{"type":"string"},"type":"array","title":"Aliases"},"origin_country":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Origin Country"},"motivation":{"items":{"type":"string"},"type":"array","title":"Motivation"},"active_since":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Active Since"},"last_seen":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Last Seen"},"target_sectors":{"items":{"type":"string"},"type":"array","title":"Target Sectors"},"target_countries":{"items":{"type":"string"},"type":"array","title":"Target Countries"},"status":{"type":"string","title":"Status"},"ttps":{"items":{"$ref":"#/components/schemas/ActorTTPOut"},"type":"array","title":"Ttps","default":[]},"tools":{"items":{"$ref":"#/components/schemas/ToolOut"},"type":"array","title":"Tools","default":[]}},"type":"object","required":["id","mitre_id","name","aliases","origin_country","motivation","active_since","last_seen","target_sectors","target_countries","status"],"title":"ActorDetailOut"},"ActorInsightOut":{"properties":{"actor_id":{"type":"string","format":"uuid","title":"Actor Id"},"payload":{"additionalProperties":true,"type":"object","title":"Payload"},"model_name":{"type":"string","title":"Model Name"},"prompt_version":{"type":"string","title":"Prompt Version"},"generated_at":{"type":"string","format":"date-time","title":"Generated At"}},"type":"object","required":["actor_id","payload","model_name","prompt_version","generated_at"],"title":"ActorInsightOut"},"ActorOut":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"mitre_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Mitre Id"},"name":{"type":"string","title":"Name"},"aliases":{"items":{"type":"string"},"type":"array","title":"Aliases"},"origin_country":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Origin Country"},"motivation":{"items":{"type":"string"},"type":"array","title":"Motivation"},"active_since":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Active Since"},"last_seen":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Last Seen"},"target_sectors":{"items":{"type":"string"},"type":"array","title":"Target Sectors"},"target_countries":{"items":{"type":"string"},"type":"array","title":"Target Countries"},"status":{"type":"string","title":"Status"}},"type":"object","required":["id","mitre_id","name","aliases","origin_country","motivation","active_since","last_seen","target_sectors","target_countries","status"],"title":"ActorOut"},"ActorTTPOut":{"properties":{"technique_id":{"type":"string","title":"Technique Id"},"technique_name":{"type":"string","title":"Technique Name"},"sub_technique_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Sub Technique Id"},"confidence":{"type":"number","title":"Confidence"},"source":{"type":"string","title":"Source"}},"type":"object","required":["technique_id","technique_name","sub_technique_id","confidence","source"],"title":"ActorTTPOut"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"RansomwareGroupOut":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"name":{"type":"string","title":"Name"},"aliases":{"items":{"type":"string"},"type":"array","title":"Aliases"},"status":{"type":"string","title":"Status"},"first_seen":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"First Seen"},"last_seen":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Last Seen"},"variants":{"items":{"type":"string"},"type":"array","title":"Variants"},"leak_site_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Leak Site Url"},"ransom_range":{"additionalProperties":true,"type":"object","title":"Ransom Range"}},"type":"object","required":["id","name","aliases","status","first_seen","last_seen","variants","leak_site_url","ransom_range"],"title":"RansomwareGroupOut"},"RansomwareVictimOut":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"group_id":{"type":"string","format":"uuid","title":"Group Id"},"victim_name":{"type":"string","title":"Victim Name"},"sector":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Sector"},"country":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Country"},"disclosed_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Disclosed At"},"source":{"type":"string","title":"Source"}},"type":"object","required":["id","group_id","victim_name","sector","country","disclosed_at","source"],"title":"RansomwareVictimOut"},"SourceHealthOut":{"properties":{"source_name":{"type":"string","title":"Source Name"},"last_success_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Last Success At"},"last_failure_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Last Failure At"},"consecutive_failures":{"type":"integer","title":"Consecutive Failures"},"status":{"type":"string","title":"Status"},"last_error":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Last Error"},"last_http_status":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Last Http Status"},"updated_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Updated At"}},"type":"object","required":["source_name","last_success_at","last_failure_at","consecutive_failures","status","last_error","last_http_status","updated_at"],"title":"SourceHealthOut"},"ToolOut":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"name":{"type":"string","title":"Name"},"aliases":{"items":{"type":"string"},"type":"array","title":"Aliases"},"type":{"type":"string","title":"Type"},"mitre_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Mitre Id"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"}},"type":"object","required":["id","name","aliases","type","mitre_id","description"],"title":"ToolOut"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}